create new domain entities
