+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/domain/Achievement.kt
@@ -1,0 +1,10 @@
+package io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain
+
+
+
+data class Achievement(
+    val title: String,
+    val description: String,
+    val points: Int,
+    val id: Int
+)
