+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/domain/User.kt
@@ -1,0 +1,12 @@
+package io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain
+
+import android.net.Uri
+import java.time.LocalDateTime
+
+data class User(
+    var displayName: String,
+    var userId: String,
+    var photoUrl: Uri?,
+    var creationDate: LocalDateTime?,
+    var lastSeenOnline: LocalDateTime?
+)
